[ot] hw/opentitan: ot_otp: move ot_otp_engine.h to public header directory
This is needed to implement other OTP top-specific implementations.
