+++ hw/opentitan/ot_otp_engine.c
@@ -35,13 +35,13 @@
 #include "hw/opentitan/ot_alert.h"
 #include "hw/opentitan/ot_fifo32.h"
 #include "hw/opentitan/ot_lc_ctrl.h"
+#include "hw/opentitan/ot_otp_engine.h"
 #include "hw/opentitan/ot_otp_impl_if.h"
 #include "hw/opentitan/ot_present.h"
 #include "hw/opentitan/ot_prng.h"
 #include "hw/opentitan/ot_pwrmgr.h"
 #include "hw/qdev-properties-system.h"
 #include "hw/qdev-properties.h"
-#include "ot_otp_engine.h"
 #include "sysemu/block-backend.h"
 #include "trace.h"
